+++ src/HashTable/hash_table_pop.cpp
@@ -17,7 +17,7 @@ enum HashTableError HashTablePopElem (hash_table_t hash_table, const char* const
 
     const size_t bucket_index = Hashing (element, strlen (element)) % kNumBucket;
 
-    const signed long long val_index =
+    const int64_t val_index =
     ListFindElem (&hash_table [bucket_index], element);
 
     if (val_index == kPoisonVal)
